+++ samples/IdentityApp/IdentityApp/Program.fs
@@ -126,7 +126,7 @@ let showErrors (errors : IdentityError seq) =
 let registerHandler : HttpHandler =
     fun (next : HttpFunc) (ctx : HttpContext) ->
         task {
-            let! model       = ctx.BindForm<RegisterModel>()
+            let! model       = ctx.BindFormAsync<RegisterModel>()
             let  user        = IdentityUser(UserName = model.UserName, Email = model.Email)
             let  userManager = ctx.GetService<UserManager<IdentityUser>>()
             let! result      = userManager.CreateAsync(user, model.Password)
@@ -142,7 +142,7 @@ let registerHandler : HttpHandler =
 let loginHandler : HttpHandler =
     fun (next : HttpFunc) (ctx : HttpContext) ->
         task {
-            let! model = ctx.BindForm<LoginModel>()
+            let! model = ctx.BindFormAsync<LoginModel>()
             let signInManager = ctx.GetService<SignInManager<IdentityUser>>()
             let! result = signInManager.PasswordSignInAsync(model.UserName, model.Password, true, false)
             match result.Succeeded with