+++ src/Giraffe/Common.fs
@@ -17,7 +17,7 @@ let inline isNotNull x = isNull x |> not
 let inline strOption (str : string) =
     if String.IsNullOrEmpty str then None else Some str
 
-let readFileAsString (filePath : string) =
+let readFileAsStringAsync (filePath : string) =
     task {
         use stream = new FileStream(filePath, FileMode.Open)
         use reader = new StreamReader(stream)